back/pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = ""
authors = ["bot <maximbotpy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
alembic = "^1.14.1"
asyncpg = "^0.30.0"
pydantic-settings = "^2.7.1"
passlib = "^1.7.4"
uvicorn = "^0.34.0"
pydantic = {extras = ["email"], version = "^2.10.6"}
fastapi-cache2 = "^0.2.2"
redis = "^5.2.1"
celery = "^5.4.0"
aiofiles = "^24.1.0"
flower = "^2.0.1"
pyjwt = "^2.10.1"
cryptography = "^44.0.2"
httpx = "^0.28.1"
itsdangerous = "^2.2.0"
authlib = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
